+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Update plugin ids for 223
 - Migrate Icons for 223
 - Update JS ModuleProvider
+- Bump versions in GitHub workflows
 
 ### Deprecated